requirements.txt
asgiref==3.6.0
autopep8==1.6.0
certifi==2022.12.7
Django==4.1.6
django-request-logging==0.7.5
djangorestframework==3.14.0
pathvalidate==2.5.2
Pillow==9.4.0
pip==22.3.1
pycodestyle==2.10.0
python-decouple==3.7
pytz==2022.7.1
setuptools==65.6.3
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.7
wheel==0.37.1
wincertstore==0.2
